Restructuration des models

+++ src/main/java/com/example/gestionEntreprise/repository/StatusRepository.java
@@ -3,4 +3,4 @@ package com.example.gestionEntreprise.repository;
 import com.example.gestionEntreprise.model.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Long> {}
+public interface StatusRepository extends JpaRepository<Status, Integer> {}
